requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
azure-identity
azure-keyvault-secrets